requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.3
asyncpg>=0.27.0
psycopg2-binary>=2.9.5
pydantic>=2.0.0
python-dotenv>=1.0.0
weasyprint>=59.0
Jinja2>=3.1.2
pyyaml>=6.0
python-dateutil>=2.8.2
pytz>=2023.3
redis>=5.0.0 